Crate tidb_query_normal_executors

This crate implements normal executors of tidb_query

Re-exports

pub use self::runner::ExecutorsRunner;

Modules

aggregate

aggregation

index_scan

limit

runner

scan

selection

table_scan

topn

topn_heap

Structs

AggCols
Row generated by aggregation.

ExprColumnRefVisitor
An expression tree visitor that extracts all column offsets in the tree.

HashAggExecutor

LimitExecutor
Retrieves rows from the source executor and only produces part of the rows.

OriginCols

ScanExecutor

ScanExecutorOptions

SelectionExecutor
Retrieves rows from the source executor and filter rows by expressions.

StreamAggExecutor

TopNExecutor
Retrieves rows from the source executor, orders rows according to expressions and produces part of the rows.

Enums

Row

Traits

Executor

Functions

get_pk

Type Definitions

IndexScanExecutor

TableScanExecutor