Struct tidb_query_normal_executors::TopNExecutor

pub struct TopNExecutor<Src: Executor> {
    order_by: OrderBy,
    related_cols_offset: Vec<usize>,
    iter: Option<IntoIter<Row>>,
    eval_ctx: Option<EvalContext>,
    eval_warnings: Option<EvalWarnings>,
    src: Src,
    limit: usize,
}

Retrieves rows from the source executor, orders rows according to expressions and produces part of the rows.

Fields
order_by: OrderByrelated_cols_offset: Vec<usize>iter: Option<IntoIter<Row>>eval_ctx: Option<EvalContext>eval_warnings: Option<EvalWarnings>src: Srclimit: usize
Implementations

impl<Src: Executor> TopNExecutor<Src>

pub fn new(meta: TopN, eval_cfg: Arc<EvalConfig>, src: Src) -> Result<Self>

fn fetch_all(&mut self) -> Result<()>